[BUGFIX] Prevent re-loading cached classes that already exist (#315)
This patch prevents a problem where it is made up to the individual cache implementation whether or not to re-load a class file when a class is already defined. Instead, making the compiler only fetch the class from cache if it is not already loaded (by checking class_exists without allowing autoloading!) prevents re-loading classes with “class already declared” errors to follow.