[FEATURE] Introduce concept of Renderables
This patch adds a new interface for "renderable" objects,
which allows any object to adopt the interface and when
assigned to the template, render itself based on the current
RenderingContext. A new argument "renderable" is added
to `f:render` to render such objects. Two implementations
are added: a StandardRenderable which renders a Node,
and a ClosureRenderable which renders arbitrary closures.

The feature is also the basis for a larger refactoring which
finally stop storing sections in the VariableProvider while
parsing, by providing a proper API for getting sections
from a parsed template object.