[FEATURE] Add getAll / addAll to ViewHelperVariableContainer
A method `getAll` is added to ViewHelperVariableContainer
which allows getting all variables for a ViewHelper as array.

A new method `addAll` is added as well, allowing a complete
set of variables to be added/overridden at the same time.

The resulting two methods allows a ViewHelper to use many
options with defaults as needed, and just one method call to
retrieve all of them and another call to update all of them.