[FEATURE] Support Vendor.Package as namespace name
Using this format (or simply Package if no vendor name is
used for the package) makes ViewHelperResolver look in
that namespace + ViewHelpers, for example:

```xml
<TYPO3Fluid.Fluid:render partial="Partial" />
```

Will internally resolve to:

```php
\TYPO3Fluid\Fluid\ViewHelpers\RenderViewHelper
```

Without the need to register the namespace. Note that the
example uses the native Fluid namespace to demonstrate,
although this namespace is always present (unless explicitly
removed by a an override).