[BUGFIX] File resolving related fixes
This patch addresses the following issues, all in TemplatePaths.

**Format-related issues**

* Method signatures contain default format which means the
  configured format (setFormat($format)) would not be respected
  and had to be passed every time. Changed to null and call to
  getFormat() when value is null.
* When extracting path arrays, default format would always be
  "html" instead of what was configured in TemplatePaths.
* Layout name prefix did not contain format when generating
  an identifier, which could cause colissions between two Layouts
  with the same name but different formats.

**General issues/improvements**

* Exception message improved to explicitly report "no paths set".
* Removed `array_values` when extracting template paths, thus
  preserving the paths' original indexes.

In addition to this, since the method signatures are changed, a
new test fixture and aliased test class for TemplatePaths is added,
which will error out if the signatures should at any point become
incompatible. The reasoning being that this particular class is
very likely to be overridden by frameworks.