[BUGFIX] Make casting of numbers the same in arguments and arrays
Problem briefly described: numbers passed in tag attributes vs. numbers passed in inline syntax (which internally uses the array
syntax parsing) handles numeric values in two different ways:

* In tags, a NumericNode is created if is_numeric is true
* In arrays, numbers are cast with (float) and matched by regexp

NumericNode also casts the value but does so by using an add-
zero trick which makes PHP do the casting based on string value,
and the input is always a string in the parser. This means that the
two different ways of passing a number will produce two different
types of variables given natural numbers as input. Passing floats
still causes the same type.

The solution is to apply the same method of casting in both cases.