[TASK] Improve performance of uncompilable templates (#335)
Problem briefly described:

Uncompilable templates are both parsed and attempted compiled on each initial rendering (once per request/context). Because no result is stored if a template is not compilable, the compiler does not know that it should not attempt to compile the template.

This patch introduces a few measures to create a compiled stub class which informs the template parser and compiler that the template was attempted compiled but determined uncompilable, so the compiler should not attempt to compile it by converting
the syntax tree to PHP code.

The stub template is achieved by allowing the template compiler to also "compile" ParsedTemplateInterface implementations which have the "compilable" flag set to false. When this is the case, the compiler creates a stub class which skips compiling when the
template is requested again.