[FEATURE] Make render method optional (#338)
ViewHelpers are no longer required to, nor suggested to, implement a render() method.

If the method does not exist, `callRenderMethod` instead calls `renderStatic` if it is implemented on the class (i.e. not inherited from the base class). If the VH class does not implement render() and inherits renderStatic() an exception is thrown, informing the developer to implement one or the other (with preference for renderStatic that is declared on the interface).