[FEATURE] Add support for ES2015 style short array syntax
This change introduces support for writing arrays in a short syntax that allows
specifying the variable names as key in a very concise way:

```
{ something, somethingElse, somethingWholeDifferent, andAlsoThis: WhichIsSomethingElse }
```

will be evaluated exactly as if it was written as

```
{ something: something, somethingElse: somethingElse, somethingWholeDifferent: somethingWholeDifferent, andAlsoThis: WhichIsSomethingElse }
```

Note though, that this currently will not work for single element arrays, since the
notation is then ambiguous with normal variable access/evaluation. Therefore, `{ foo }`
will just mean "the value of variable foo".