[BUGFIX] Prevent evaluating ExpressionNode unless demanded
Adds a new interface which must be used by expression node
types in order to signal that the expression node must be
evaluated during parsing of the template.

History:

Parse-time evaluation of ExpressionNodes is a remnant from
the time when there was a separate legacy namespace
extraction based on expression nodes. While there currently
are expression-like template processors, there are no known
implementations out there which must have the node evaluated
during parse time.

Choices:

Rather than simply remove the evaluation altogether, the new
interface is added as a quicker way to migrate *if* there are
private repositories out there implementing these nodes.
If the evaluation was removed entirely, the only alternative left
would be migrating to a template pre-processor which is by
nature quite different from the expression nodes. So, as a
convenience, this interface is left in place to be implemented
as substitute for the old interface, requiring no further change.

Close: #300