[FEATURE] Deprecate evaluateCondition() in favor of verdict()
The AbstractConditionViewHelper method evaluateCondition
was less than ideal, in that it:

* Only received arguments and thus could not support context
  related conditions.
* Was protected which prevents a future refactoring to call the
  method to evaluate directly from the compiled code.
* Was not strictly requiring an array of arguments which meant
  technically unsafe code (static analysis would demand guard)

Instead of breaking the API, the old method is deprecated in
favor of a new method which strictly requires arguments as an
array and the current instance of the rendering context.

The new method is made public though it is still not being
called as such - this is in preparation for a future improvement
of the compiled code to clean up the current closure-based
solution for evaluating the if-elseif-else structure.

In addition, this patch:

* Refactors the condition ViewHelper so using the base class
  does not automatically register a “condition” argument.
* Refactors `f:if` to use the new API method.