[BUGFIX] Process paths individually in warmup
A bug in the warmup code meant that all templates,
partials and layouts which were in sub-folders in
the root paths, would cause file not resolved errors
due to using basename() on the absolute file path.

To remedy this, each path has to be traversed by
itself and the relative file path has to be determined
by trimming off the root path. The result is that any
level of sub-folders can be supported in warmup.