[FEATURE] Improve inline syntax for arguments and passing
This patch improves the inline syntax of Fluid in the
following ways:

* The “pipe” character can be used instead of “->” to
  pass tag contents, e.g. {variable | f:format.raw()}. This
  is done to approach shell syntax and for easier picking
  up of Fluid syntax if you come from Twig.
* Arguments for inline VH calls can now be specified with
  tag attribute syntax. Commas were already optional which
  allowed syntax like “{v:h(foo: bar baz: 123)}”. This is now
  homogenised completely so you can use expressions like
  “{v:h(foo=“{bar}” baz=“123”)}” - which means that you
  can now migrate a tag-based usage to inline syntax
  without having to touch the arguments you provided as
  tag attributes.

All is achieved by adding two characters to the list that is
detected as inline syntax - “|” and “=“ - and by turning
the matching of “->” into a multiple match of either “->”
or the new “|”.

All legacy syntax is untouched. Tag syntax is untouched.

Full disclosure: a similar change was previously rejected
by myself: https://github.com/TYPO3/Fluid/pull/343. The
only arguments I can give for accepting this current patch
but rejecting #343, are:

* This patch modifies far less of the regular expressions
  and sticks to adding characters and converting a character
  match to a list of options.
* It is matching the current research I am doing on using a
  lexer to parse Fluid (a project which also tries to include
  the ES2015 style array syntax but can do so without any
  concerns about bloating of even overflowing the regexp).