[FEATURE] Allow content argument on ViewHelpers without trait
This change deprecates CompileWithContentArgumentAndRenderStatic
in favor of a similar, but explicit, solution on AbstractViewHelper.

Content argument is possible on any ViewHelper after this
change but must explicitly be opted-in by one of two methods:

* Set `protected $contentArgumentName` to name of argument.
* Override `resolveContentArgumentName` and return name.

The latter of the two methods is implemented so ViewHelpers
can be plug-and-play compatible even with the trait hollowed
out and deprecated.

A couple of ViewHelpers and related tests have additionally
been refactored to not assume that render() is implemented
(the method is now optional), instead using static calling. This
allowed removing the deprecated trait from those ViewHelpers.

Close: #400