[TASK] Refactor child node evaluation to iterable
This patch slightly refactors the way evaluateChildNodes
behaves, allowing the return value from getChildNodes
to be any type of iterable type/object.

There are currently no use cases in existence that would
change a node to return non-arrays from getChildNodes
but in order to support this in the future we can refactor
now without ill effects. For example this would allow the
TemplateParser to be refactored to use Generator for
yielding parsed nodes.