[FEATURE] Inline fluid ViewHelper
Allows rendering a variable containing Fluid code as
if the code contained in the variable were part of the
template invoking the ViewHelper.

    $view->assign('variable', 'value of my variable');
    $view->assign('code', 'My variable: {variable}');

And in the template:

    {code -> f:inline()}

Which outputs:

    My variable: value of my variable

You can use this to pass smaller and dynamic pieces of Fluid code
to templates, as an alternative to creating new partial templates.