[FEATURE] Allow variables assigned with dotted path
Enables assignment of template variables with a dotted
path as name:

    $view->assign(‘parent.property’, ‘newValue’);

Each value in the path leading up to the final property
name is created as an array if it does not exist, and is
read by reference until a subject is identified. Then the
value is set on that subject.

If a path points to an object and the path to the property
contains one or more scalar values, assignment is
refused with an exception clearly stating why a subject
could not be resolved.

The feature is also enabled for the JsonVariableProvider
by making it call the StandardVariableProvider’s method.