[TASK] Improve performance of hardcoded booleans and empty executions
Improves performance of output escaping when the
output that would have been escaped is empty, as
determined by the compiled state of the output (empty
execution instruction means empty output).

Improves performance throughout when boolean arguments
are used with hardcoded values like ‘1’ for TRUE, which is
very frequently the case. Compiles these cases as direct
value evaluations and casting to boolean, and catches
cases where only a single node needs to be evaluated so
the node’s return value is immediately cast as boolean.

Avoids the relatively expensive 1) conversion of a one-value
array to compiled array and 2) invoking the boolean parser
to consume this array and cast the value. And makes the
compiled template source code file significantly smaller.