[TASK] Remove all traces of caching
Removes all template compiling and -caching from Fluid.

This includes:

* TemplateCompiler, NodeConverter
* FluidCacheInterface and all implementations
* All compile() methods on ViewHelpers and interface

The performance detriment ranges from 10% to 150%
for some large and complex templates. However, it will
be possible to further improve the performance of the
Sequencer to reduce this.

The major benefit is an extreme reduction of the total
complexity of writing a well-performing ViewHelper,
and reduces the ways a ViewHelper can be rendered
to one instead of the current two different ways.

Further reduction of the API inside ViewHelpers will
provide an additional performance boost, as will the
plan to replace the use of 1) template pre-processors
and 2) expression nodes by 1) integrating the features
currently handled by pre-processors into the Sequencer
and 2) allowing a ViewHelper to behave as expression
if so registered.

Signed-off-by: Claus Due <claus@namelesscoder.net>