[TASK] Reduce complexity of ViewHelpers
First step to reduce internal API of ViewHelpers:

* Removes postParseEvent
* Removes ViewHelperNode
* Removes ViewHelperInvoker (note below)

The removal of the ViewHelperInvoker may seem drastic
but has to be viewed in a historical context - the original
purpose of this class was to ensure that frameworks which
wanted to allow render method arguments, could do so
via overriding this special invocation class.

Since render methods are no longer supported in any way
the invocation class can be removed. At first this will
consolidate the ViewHelper invocation entirely in the
ViewHelper class - in the longer term it makes it easier
to refactor argument handling to the up-and-coming
component argument pragma.

Signed-off-by: Claus Due <claus@namelesscoder.net>