[TASK] Improve array sequencing
Prevents some memcopy, business logic and/or
return type ambiguity that would otherwise be
required to sequence an array and then use the
result as various object types.

Converts ArrayNode to an ArrayAccess and uses
a parameter passed by reference when sequencing,
then assigns the parsed values directly to it.

Signed-off-by: Claus Due <claus@namelesscoder.net>