[TASK] Remove Interceptors
Replaces the concept of Interceptors with internal
and much more efficient escaping-specific handling.

Sacrifices a parse-time flexibility which is rarely
used, instead preferring ViewHelpers or other
components to be used instead of interceptors.

Removes interception of ordinary text which avoids
calling interception methods for anything quoted
and any sub-expression parts.

Signed-off-by: Claus Due <claus@namelesscoder.net>