[FEATURE] Replace View with FluidRenderer, remove MVC
Introduces a file- or source-based entry point for
rendering Fluid templates and completely deprecates
the old View scope including TemplatePaths.

This also means the MVC opinionated nature of Fluid
is changed to a simpler concept of providing the
starting file and the collection of paths where other
files that may be sub-rendered can exist.

A replacement for the legacy paths handling will then
be created.

Signed-off-by: Claus Due <claus@namelesscoder.net>