[FEATURE] Add Atoms feature
An "Atom" in Fluid is essentially a template file which
exists in a path. The path can be added to Fluid as a
namespace, much like how PHP namespaces are
added to Fluid so an expected class name can be
determined without scanning through files.

When a path is added under a namespace, using that
namespace in tag- or inline syntax implies that you
want to render (actually, import the node tree of) the
template identified by the tag name.

For example:

    Resources/Private/Atoms/myAtom.html

Registered as namespace "foo" becomes in Fluid:

    <foo.myAtom />

Furthermore, f:parameter is added which allows a
developer to embed ArgumentDefinition into template
files - these arguments are then enforced when the
Atom is rendered (making parsing fail if a required
argument is not provided to the custom tag/inline).

Arguments passed automatically become variables:

    <foo:myAtom arg1="bar" arg2="something" />

Become {arg1} and {arg2} inside the template. Any
arbitrary argument is allowed - but defining a required
argument means at least that argument has to be
passed to the Atom.

Declaring a parameter as "bool" type also means that
the argument is parsed as a boolean node, meaning it
enables the special boolean syntax for that argument:

    <foo:myAtom boolArgument="{foo} == 4" />

Which means {boolArgument} is either "true" or "false"
when it reaches the Atom.

Signed-off-by: Claus Due <claus@namelesscoder.net>