[FEATURE] Add SequencingComponentInterface and implementations
This change adds a new interface, SequencingComponentInterface,
which lets a Component take over the sequencing that is
usually done by the Sequencer.

Also adds implementations which allow unparsed Fluid code
inside special tags:

* f:comment is rewritten to ignore any nested Fluid.
* f:description is added and allows unparsed Fluid
  tag content that describes a template or section.
* f:example is added which allows embedding Fluid
  example code for a template or section.

These special Components are not rendered but are
instead embedded and can be extracted from a
parsed Fluid template. When used with an open and
close tag along with tag content, the tag content is
not parsed as Fluid and becomes a plain string.

Signed-off-by: Claus Due <claus@namelesscoder.net>