[BUGFIX] Adjust behavior of f:extends
Makes the f:extend ViewHelper render the Atom's child
nodes on execution. Fixes a problem where extending an
Atom which contains a section, and then declaring a
section with the same name in the importing template,
would cause the section from the Atom to render instead
of the section from the importing template.