[BUGFIX] Restore support for data- prefixed arguments
Due to a regression in https://github.com/TYPO3/Fluid/pull/419,
a TagBasedViewHelper which was called with one or more
data- prefixed arguments would discard the prefixed arguments
and the output would not contain the attribute.

This patch fixes that regression by moving the handling of
unregistered arguments to the initialize() function in order
to make sure the additional attributes are handled *after*
the reset() method has been called on TagBuilder.

References: #419