Add Argument viewhelper
Add a new `f:argument` viewhelper to be able to add
arguments to a node by using `f:argument` as child node.

Usage example:
```
<my:node>
	<f:argument name="content">Hello World</f:argument>
</my:node>
```

The sequencer does now put validation errors into a stack
so we are able to revalidate a node before it is closed but
children have been parsed. So we can still show the initial
problem if e.g. a required property was not defined.

Resolves #427