[BUGFIX] Drop circular dependency RenderingContext<->AbstractTemplateView
The constructor of AbstractTemplateView instantiates
RenderingContext, giving itself over as constructor
argument.
However, there is no point the RenderingContext knows
the view, and $this->view is actually completly unused.
There is also no getter or anything accessing the
property.
The patch resolves this circular dependency by dropping
the required constructor argument in RenderingContext.