[FEATURE] Show useful error message on string to array conversion
If an array is used in a template as variable, the EscapingNode
tries to cast it to a string which does not work but throws an error.
To make the error message easier to understand this case is now
catched and an Exception with the content of the array is shown
instead of only the stacktrace.