[BUGFIX] Avoid fatal error during inheritance in CountableIterator
PHP8.1 added strict return types to internal methods, functions and
interfaces which emits E_DEPRECATION warnings if not matching or used
correctly. For interfaces this ends in fatal error due inheritance.

This commit adds the introduced #[\ReturnTypeWillChange] attribute
to the methods of 'CountableIterator' fixture to avoid this warning
and fatal error with PHP8.1 but keeps backwards compatibility.