[BUGFIX] Ensure TemplateCompiler->sanitizeIdentifier return a string
Calling class_exists(null) emits a "Passing null to parameter" PHP
E_DEPRECATED warning since PHP8.1, which occured in TemplateCompilerTest
when executed with PHP8.1 and E_ALL error reporting enabled.

Because of an inproper mock of `TemplateCompiler->sanitizeIdentifier()`
in `TemplateCompilerTest` return null instead of an string, this error
occured during testing with PHP8.1.

This commit ensures that `TemplateCompiler->sanitizeIdentifier` returns
always a string, string typecast is added on the return of preg_replace(),
thus avoiding the E_DEPRECATED warning in `TemplateCompiler->has()` as
class_exists got not called with null for normal usage. Furthermore the
mocking is corrected to return an empty string instead of null for one
testcase, which match the expected return, thus avoiding this error there.