[BUGFIX] Allow action-only templates to run
The current templatePaths functionality is highly limited,
in order to run Fluid with an "ADR" approach where each
Action represents a single PHP class (single responsibility),
Fluid always expected a Controller in order to build the path.

This change addresses the issue when a TemplatePaths object
is built with an empty controller (""), to resolve the paths
correctly when trying to locate the proper file path.