[TASK] Change property docblock of $contentArgumentName how to set it (#583)
Trait 'CompileWithContentArgumentAndRenderStatic' provides
magic handling retrieving to retrieve the first optional
registered argument as render children closure, if argument
is specified. Additional this can be explicitly set with
'contentArgumentName', which is also recommended in the
trait itself.

However this cannot be done with adding the property to
the ViewHelper itself, as this violates how trait works
and ends in a fatal error would be emited. To mitigate
this, either override the 'resolveContentArgumentName()'
method and return the name or through assigning the name
to this property in method ViewHelper constructor
'__construct()', using for example the well known method
'initializeArguments()' do not work here.

This commits changes the comment of that property in the
trait with this example and removes the not working hint
to copy and set the property directly in the consuming
ViewHelper.

Tests are adjusted and extended to cover only the working
constellations, thus solving issue #582 through clarifing
how to set the argument correctly.