[TASK] Remove coveralls code coverage stats
Coverage is a lie. With current configuration, it
does not really reflect actual test statistics and
"97%" is simply misleading since it ignores files
that are not loaded at all durnig tests.

As such, this statistic does not actively help us
in current development and should be dropped.

Naturally, patches that have parser impact need
proper test coverage and won't be accepted without.
A misleading coverage generator does not help
us here.