[TASK] Refactor functional tests
A lot of functional tests extend BaseConditionalFunctionalTestCase
or BaseFunctionalTestCase. These try to set up a generic functional
test system by providing API for both a "non-cached" and "cached"
run for functional tests. The weird class and inheritance construct
leads to a lot of skipped tests that clutter phpunit output.

That setup is hard to follow and becomes a victim of its own
complexity since it does not test cached behavior in the end:
The abstract method 'getView()' simply ingores '$withCache' argument,
so 'cached' behavior' has never been tested.

The patch simplifies the construct and takes care cache-hit
scenarios are actually tested, basically by executing the template
parsing twice for singe tests. A new AbstractFunctionalTestCase is
provided as base and the old construct is winded up.

The new setup shows that especially ternary condition behave
diffently when cached. According tests are commented and marked
with @todo annotations for now since they outline bugs in
current implementation.

All in all, this patch simplifies functional tests setup a lot
and makes them more easy to follow and understand, and hopefully
to fix as well.