[BUGFIX] Correct cached ternary expressions
Ternary expressions are broken with cached templates.
Even with simple things like this:

$view->assign('myVariable', false);
{myVariable ? 'yes' : 'no'}

This evaluates correctly to 'no' on first call with cold
caches, but to 'yes' when a cached template is used
on subsequent calls.

This came up when the functional tests have recently been
fixed to correctly test cached behavior.

The fix is to evaluate() the "left" / "check" side just
like the non-cached variant does.