[TASK] Refactor GroupedFor VH tests to functionals (#625)
The heavy mocking of the ViewHelper unit tests can be
misleading. It's much more obvious when test sets simply
feed appropriate fluid strings to the parser and check
rendered output. We did a similar thing to the TYPO3 CMS
VH's with great success already.

The patch turns all unit tests of GroupedForViewHelper
into functional tests, removes a dupe and applies
some overall refactoring.

With the now checked output, one test scenario now
reveals a bug in the VH implementation, the data set
is disabled for now and marked with todo.

As side effect, this VH now tests cached scenarios. This
is also a preparation of PR #578 and relates to #211.