[BUGFIX] ParserRuntimeOnly trait compile() returns string
ViewHelperInterface->compile() is annotated to return
string. ParserRuntimeOnly violates this and returns
null instead. This leads to different behavior in
consuming ViewHelpers between non-compiled and compiled
versions.

The change adapts ParserRuntimeOnly->compile() to
return an empty string and changes NodeConverter to
deal with it. A string cast is added to stay compatible
in case some other compile() method returns null as well.