[TASK] Improve Fluid executable
The executable is a php application, but due to the missing
file extension no code support is available.
Also, all code resides inside the executable file, excluding it
from supporting tools like php-cs-fixer or phpstan.

Dividing the executable file into a real class, a caller script
and a very minimal executable script exposes the PHP
code to supportive tools and IDE functions.

This is a preparation patch to improve the executable files
functionality and provide better user experience with it.