[TASK] Fix tests using willReturnOnConsecutiveCalls() (#680)
We switched from withConsecutive() to
willReturnOnConsecutiveCalls() when adding
phpunit 10 support. This was partially broken
since various tests were designed to test the
input, which withConsecutive() does, but
willReturnOnConsecutiveCalls() does not.

Switch to a callback in some places, remove
a couple of heavily mocking tests altogether
and change some tests to functional tests.

In general, we should continue to avoid various
heavy mocking tests and various whitebox tests,
and substitute them with good functional tests.