[BUGFIX] Allow using undefined variables in InlineViewHelper (#665)
InlineViewHelper crashes with a children closure that evaluates to
NULL, since parse() requires a non-null string as argument.

This is a bit edge-casy, but may happen when a variable is
not set or its value is null. The patch sanitizes this case and
adds a couple of tests to verify.