[TASK] Support Neos/Flow class compiler
With 2.7.3, we extracted all logic from bin/fluid
to src/Tools/ConsoleRunner.php and src/Tools/console.php,
where ConsoleRunner contains the worker class and
console.php the autoload and instantiation logic
as global code.

This works, but one could argue it's ugly having
code within a source directory that is supposed to
contain classes only.

More importantly, it turns out Neos/Flow can not deal
with global code that just "runs" in a folder that
expects to contain all code being encapsulated in
classes only. The proxy mechanism of Neos/Flow stumbles:

Undefined variable $argv in Packages/Libraries/typo3fluid/fluid/src/Tools/console.php line 22

The patch moves the global scope code back into bin/fluid.

Additionally, the Heredoc in ConsoleRunner gets a
minor update because the Neos/Flow class compiler
interpretes 'name' of the phrase 'class name' as a
class name which leads to another exception:

The name of the class "name" is not the same as the filename which is "ConsoleRunner.php".

Resolves: #691