[FEATURE] Add runtime cache for object access in fluid
This change reduces PHP function calls when accessing
multiple instances of the same class in a fluid template. The current
implementation checks the class structure with each property access
for the availability of getter or asserter methods, which can be
cached after the first check and re-used for all subsequent accesses
of the same property.

Due to the way variable providers are implemented, this cache can also
be re-used in partials, layouts and sections.