[TASK] Remove variable accessors from ObjectAccessorNode
In an attempt to improve compiled template performance,
82911b85 and 735f3362 established a second constructor
argument on ObjectAccessorNode, handing over "accessors".
This was later reverted in 48e43466 to fix #255.

As a result, ObjectAccessorNode is never called with
a second constructor argument in business code, so
$this->accessors is always an empty array.

The patch removes the second constructor argument
again and simplifies the generated template code
in NodeConverter accordingly.

Related: #255