[TASK] Avoid and deprecate ParserRuntimeOnly trait (#744)
This trait has been added with 127d8e70 for
no apparent reason: It was used three times
with one consumer overriding the render()
method again. There is no point to have a
trait for methods that only return null
or empty strings.

To simplify the construct a bit, the trait
is avoided and marked as @deprecated. Neither
Flow nor TYPO3 use the trait in own VH's,
impact should be quite small.