[TASK] Avoid and deprecate ViewHelperCompiler (#745)
The ViewHelperCompiler helper class has been
introduced a while ago and shrinked over
time again. It consists of one method with
an sprintf() and two constants nowadays.

It is used in the two traits
CompileWithRenderStatic and
CompileWithContentArgumentAndRenderStatic.

To simplify the code a bit, the code of
ViewHelperCompiler is now inlined to the
two traits, which makes them easier to
read and understand. ViewHelperCompiler
is marked @deprecated, neither Flow nor
TYPO3 use that class, so impact should be
rather low.