[TASK] Avoid and deprecate getAccessibleMock() (#758)
This is another test related method that tends
to perform whitebox tests. The patch refactors
usages to avoid the method which uses some phpunit
methods that are soft-deprecated itself. The
strategy is to get rid of a lot of whitebox
tests, but to fetch and set state using PHP
reflection directly in cases where this is not
useful.