[TASK] Avoid RenderingContextFixture
Lots of test subjects carry the rendering
context around and access it to perform
things.

The "god" test fixture RenderingContextFixture
is used to mock-away this dependency in tests.
This is however hard to understand and adds
complexity to tests.

The patch uses RenderingContext directly in
most cases, or switches to mocks of the
RenderingContextInterface.

Strictly speaking, this makes some "unit" tests
de facto integration / functional tests, but
moving tests around or refactoring to mock
RenderingContext again is a task for further
patches.

For now, this patch should make various tests
easier to understand and follow.