[BUGFIX] Correct compiled f:if VH edge cases
This fixes a bunch of f:if ViewHelper edge cases.

Most test cases have been established with previous
patches already, inactive ones with issues on compiled
templates are activated now.

Especially else and else if code edge cases are in
line with non-compiled execution now.

The patch essentially rewrites the generated compiled
PHP code, as well as its execution flow within
AbstractConditionViewHelper. The internal changes
to "__" prefixed arguments are not considered
breaking, TYPO3, Flow and other consumers should in
general not suffer from this and can enjoy more solid
behavior.

The code is better documented and should be
more easy to follow in general.

The patch adds just one further commented test set,
but this is related to not-compiled templates and
will be sorted out with another patch.

The tests show that the f:if VH will now return
empty strings instead of null more often. This
is good and comes as additional benefit in order
to further harden method return types later.

Resolves: #769